Convert `@ProduceStyle()` to amendment
