+++ package.json
@@ -49,6 +49,7 @@
     "@frontmeans/drek": "^1.1.0",
     "@frontmeans/render-scheduler": "^1.6.0",
     "@hatsy/http-header-value": "^3.5.3",
+    "@proc7ts/amend": "^1.0.0-dev.1",
     "@proc7ts/primitives": "^2.1.0",
     "@proc7ts/push-iterator": "^2.5.0",
     "@proc7ts/supply": "^1.0.0"
