+++ src/styp/index.ts
@@ -7,6 +7,6 @@ export * from './component-styp-dom.format';
 export * from './component-styp-dom.format-config';
 export * from './component-styp-format';
 export * from './component-styp-renderer';
-export * from './produce-style.decorator';
+export * from './produce-style.amendment';
 export * from './style-producer-support.feature';
 export * from '../theme';
